% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_julian.R
\name{from.julian}
\alias{from.julian}
\title{Convert From Julian Date}
\usage{
from.julian(j.date)
}
\arguments{
\item{j.date}{Text string in the format of YYYYDDD}
}
\value{
object of class Date
}
\description{
This function takes a Julian date and converts it to a date object.
The Julian format expected is a four digit year and the three digit number of the day of the year.
For example January 1st's number is 001 and December 31st is either 365 or 366 if it is a leap year.
}
\examples{
from.julian("2017090")
}
\keyword{date}
\keyword{julian,}
