% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iseriesRODBC.R
\name{iseriesRODBC}
\alias{iseriesRODBC}
\title{Connect to iSeries Server}
\usage{
iseriesRODBC(SystemAddress = "SERVICE", Library = NA, User, Password)
}
\arguments{
\item{SystemAddress}{IP or host name of server. Defaults to SERVICE for my convienence.}

\item{Library}{Specifiy and optional library to restrict the connection to.}

\item{User}{username to access server}

\item{Password}{password to access server}
}
\value{
object of class RODBC
}
\description{
Creates a connection device to using an ODBC iSeries driver and odbcDriverConnect from the
    RODBC package.
}
\seealso{
\code{\link[RODBC]{odbcConnect}}
}
