% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_julian.R
\name{to.julian}
\alias{to.julian}
\title{Convert To Julian Date}
\usage{
to.julian(my.date)
}
\arguments{
\item{my.date}{Date object}
}
\value{
text string in the format of YYYYDDD
}
\description{
This function takes a Date object and converts it to a text string in Julian date format.
The Julian formatis a four digit year and the three digit number of the day of the year.
For example January 1st's number is 001 and December 31st is either 365 or 366 if it is a leap year.
}
\examples{
to.julian(as.Date("2017-03-31"))
}
\keyword{date}
\keyword{julian,}
